requirements.txt
asgiref==3.8.1
coverage==7.5.4
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
flake8==7.1.0
gunicorn==22.0.0
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
psycopg2==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
PyJWT==2.8.0
python-dotenv==1.0.1
rest-framework-simplejwt==0.0.2
sqlparse==0.5.0
